Refactor code using objects and fix bugs
Refactored the code to use a player object with a name and score to
allow for easier editing and expansion of game features in the future.
Fixed several bugs: game now resets when attempting to continue to play
after a winner has been declared. Added missing text for before first
round was played. Removed bloat/unnecessary code now that we are using
objects to define the player class.